add results

+++ benchmark/linear-interpolation/Python/multithread/python_multi_thread.py
@@ -6,9 +6,9 @@ MAX_THREADS = os.cpu_count()
 def save(arr):
     f = open("resultPython.txt", "w")
 
-	cuantity = 1000000
+    cuantity = 1000000
 
-	limitedSelection = len(arr) / cuantity
+    limitedSelection = int(len(arr) / cuantity)
 
     for i in range(0, cuantity):
         f.write(f"{i * limitedSelection} {arr[i * limitedSelection]}\n")
